Extract AiLodge to eliminate duplication across Anthropic/Gemini/PocAi lodges
