+++ src/main/kotlin/werewolf/lodge/AnthropicLodge.kt
@@ -1,27 +1,12 @@
 package werewolf.lodge
 
-import werewolf.ai.AiPersonalities
-import werewolf.ai.AiPlayer
-import werewolf.ai.Instruction
+import werewolf.ai.LanguageModel
 import werewolf.ai.anthropic.AnthropicLanguageModel
-import werewolf.game.Player
-import werewolf.game.Role
 
-class AnthropicLodge(humanConnection: HumanConnection) : Lodge(humanConnection) {
-    override fun assignments(): List<Pair<Player, Role>> {
-        val roles = listOf(
-            Role.HUNTER, Role.VILLAGER, Role.VILLAGER, Role.VILLAGER, Role.MADMAN,
-            Role.WEREWOLF, Role.SEER, Role.MEDIUM, Role.WEREWOLF,
-        ).shuffled()
-        val personalities = AiPersonalities.list.shuffled()
-        val aiPlayers = roles.drop(1).mapIndexed { i, role ->
-            AiPlayer(role, AI_NAMES[i], AnthropicLanguageModel(HAIKU_MODEL), Instruction(AI_NAMES[i], personalities[i])) to role
-        }
-        return listOf(createPlayer(roles[0], "Ivan") to roles[0]) + aiPlayers
-    }
+class AnthropicLodge(humanConnection: HumanConnection) : AiLodge(humanConnection) {
+    override fun createLanguageModel(): LanguageModel = AnthropicLanguageModel(HAIKU_MODEL)
 
     companion object {
         private const val HAIKU_MODEL = "claude-haiku-4-5-20251001"
-        private val AI_NAMES = listOf("Alice", "Bob", "Charlie", "Dave", "Eve", "Frank", "Grace", "Heidi")
     }
 }
